requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
elevenlabs==1.57.0
Flask==3.1.0
groq==0.24.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
playsound==1.3.0
PyAudio==0.2.14
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
SpeechRecognition==3.14.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
-e .